Add files via upload
This app prompts the user with a chess board in order for them to solve the classic 8 Queens problem known throughout the chess community. 

Included with the chess board are a few features including the clear option to restart, a hint checkbox to show the available spaces, and a right click feature which will remove a single queen.  

https://en.wikipedia.org/wiki/Eight_queens_puzzle#:~:text=The%20eight%20queens%20puzzle%20is,in%20the%20mid%2D19th%20century.